CRM Back end/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.4,<5.5
django-cors-headers>=4.6,<5.0
django-filter>=24.3,<25.0
django-environ>=0.12,<1.0
psycopg[binary]>=3.2,<4.0
redis>=5.2,<6.0
celery>=5.4,<6.0
django-celery-beat>=2.7,<3.0
django-celery-results>=2.5,<3.0
Pillow>=11.0,<12.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
drf-spectacular>=0.28,<1.0
django-import-export>=4.2,<5.0
pyotp>=2.9,<3.0
qrcode[pil]>=8.0,<9.0
cryptography>=43.0,<44.0
bleach>=6.2,<7.0
openai>=1.50,<2.0
anthropic>=0.39,<1.0
